feat: create kubernetes client when app is started
